ДЗ 3й недели
Учел комментарии наставника.

Вопросы:
1. Не разобрался с тестированием. Смотрим файл "__test__/Auth.tests.js". При запуске получаю ошибку ниже

Error: Could not find "store" in the context of "Connect(App)". Either wrap the root component in a <Provider>, or pass a custom React context provider to <Provider> and the corresponding React context consumer to Connect(App) in connect options.

При попытке выполнить render(<Provider store={store}><App/></Provider>); такая же ошибка.

2. Не разобрался с тестированием middleware. Смотрим файл "__test__/authMiddleware.spec.js"

Не понял как задать выходной параметр функции serverLogin, чтобы функция возвращала объект, а не число/строку/булевое значение. В текущей реализации получаю:

Cannot read properties of undefined (reading 'success')
