+++ src/__test__/Login.test.js
@@ -1,6 +1,6 @@
 import React from "react";
 import { render, fireEvent } from "@testing-library/react";
-import App from "./App";
+import App from "../App";
 
 describe("toggle Auth form", () => {
   it('click link and toggle form', () => {
