+++ src/milldewares/cardMiddleware.js
@@ -1,28 +1,10 @@
-import {LOG_OUT, AUTH, CHANGE_CARD, errorLogIn, logIn, errorChangeCard, LOAD_CARD, setCard} from "./actions";
-import {serverLogin, serverSaveCard, serverLoadCard} from "./loftTaxiServer";
+import {CHANGE_CARD, errorChangeCard, LOAD_CARD, setCard} from "../actions";
+import {serverSaveCard, serverLoadCard} from "../loftTaxiServer";
 
-export const authMiddleware = (store) => (next) => async (action) => {
+export const cardMiddleware = (store) => (next) => async (action) => {
   let resp;
 
   switch (action.type) {
-    case AUTH:
-      const {email, password} = action.payload;
-
-      resp = await serverLogin(email, password);
-
-      if (resp.success) {
-        store.dispatch(logIn(resp.token));
-
-        localStorage.setItem('isLoggedIn', '1');
-        localStorage.setItem('token', resp.token);
-      } else {
-        store.dispatch(errorLogIn(resp.error));
-      }
-      break;
-    case LOG_OUT:
-      localStorage.clear();
-      next(action);
-      break;
     case CHANGE_CARD:
       const payload = action.payload;
       resp = await serverSaveCard({...payload, token: store.getState().auth.token});
