+++ src/store.js
@@ -1,5 +1,6 @@
 import {createStore, applyMiddleware} from "redux";
 import rootReducer from './reducers'
-import {authMiddleware} from "./authMiddleware";
+import {authMiddleware} from "./milldewares/authMiddleware";
+import {cardMiddleware} from "./milldewares/cardMiddleware";
 
-export const store = createStore(rootReducer, applyMiddleware(authMiddleware));
+export const store = createStore(rootReducer, applyMiddleware(authMiddleware, cardMiddleware));
